AccesoPbRol Controller

+++ src/main/java/com/banquito/core/banking/seguridadbanco/domain/AccesoPbRolPK.java
@@ -1,6 +1,7 @@
 package com.banquito.core.banking.seguridadbanco.domain;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
@@ -12,15 +13,15 @@ import lombok.Setter;
 @Embeddable
 public class AccesoPbRolPK implements Serializable{
     @Column(name="COD_ROL", nullable=false)
-    private Integer codRol;
+    private BigDecimal codRol;
     @Column(name="COD_PER_BAN", nullable = false)
-    private Integer codPerBan;
+    private BigDecimal codPerBan;
 
     
     public AccesoPbRolPK() {
     }
 
-    public AccesoPbRolPK(Integer codRol, Integer codPerBan) {
+    public AccesoPbRolPK(BigDecimal codRol, BigDecimal codPerBan) {
         this.codRol = codRol;
         this.codPerBan = codPerBan;
     }
